docs: add config as comments within documentations examples
The Javascript documentation has at the start of the code sample a
comment line that activates the rule with a specific config. Tried to do
it for the Typescript documentation. Also change a bit of the text in
the front of some correct/incorrect example to be more like the
Javascript one.